+++ ont_fast5_api/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-__version__ = '2.0.0'
+__version__ = '2.0.1'
 __version_info__ = tuple([int(num) for num in __version__.split('.')])
 
 # Set up a default NullHandler in case we don't end up using another one