+++ ont_fast5_api/conversion_tools/multi_to_single_fast5.py
@@ -8,7 +8,8 @@ import os
 
 from ont_fast5_api import __version__
 from ont_fast5_api.conversion_tools.conversion_utils import get_fast5_file_list, get_progress_bar
-from ont_fast5_api.fast5_file import EmptyFast5
+from ont_fast5_api.fast5_file import EmptyFast5, Fast5FileTypeError
+from ont_fast5_api.fast5_interface import check_file_type, MULTI_READ
 from ont_fast5_api.multi_fast5 import MultiFast5File
 
 logging.basicConfig(level=logging.INFO)
@@ -16,17 +17,16 @@ logger = logging.getLogger(__name__)
 exc_info = False
 
 
-def batch_convert_multi_files_to_single(input_path, output_folder, threads, recursive):
-
+def batch_convert_multi_files_to_single(input_path, output_folder, threads, recursive, follow_symlinks):
     pool = Pool(threads)
-    file_list = get_fast5_file_list(input_path, recursive)
+    file_list = get_fast5_file_list(input_path, recursive, follow_symlinks=follow_symlinks)
     pbar = get_progress_bar(len(file_list))
 
-    def update(results):
-        output_file = os.path.basename(results.popleft())
+    def update(result):
+        input_file = result[0]
         with open(os.path.join(output_folder, "filename_mapping.txt"), 'a') as output_table:
-            for filename in results:
-                output_table.write("{}\t{}\n".format(output_file, filename))
+            for filename in result[1]:
+                output_table.write("{}\t{}\n".format(input_file, filename))
         pbar.update(pbar.currval + 1)
 
     if not os.path.exists(output_folder):
@@ -45,22 +45,31 @@ def batch_convert_multi_files_to_single(input_path, output_folder, threads, recu
 
 
 def convert_multi_to_single(input_file, output_folder, subfolder):
-    results = deque([os.path.basename(input_file)])
+    output_files = ()
     try:
-        with MultiFast5File(input_file, 'r') as multi_f5:
-            for read in multi_f5.get_reads():
-                try:
-                    output_file = os.path.join(output_folder, subfolder, "{}.fast5".format(read.read_id))
-                    create_single_f5(output_file, read)
-                    results.append(os.path.basename(output_file))
-                except Exception as e:
-                    logger.error("{}\n\tFailed to copy read '{}' from {}"
-                                 "".format(str(e), read.read_id, input_file), exc_info=exc_info)
+        output_files = try_multi_to_single_conversion(input_file, output_folder, subfolder)
     except Exception as e:
         logger.error("{}\n\tFailed to copy files from: {}"
                      "".format(e, input_file), exc_info=exc_info)
-    finally:
-        return results
+    return input_file, output_files
+
+
+def try_multi_to_single_conversion(input_file, output_folder, subfolder):
+    output_files = []
+    with MultiFast5File(input_file, 'r') as multi_f5:
+        file_type = check_file_type(multi_f5)
+        if file_type != MULTI_READ:
+            raise Fast5FileTypeError("Could not convert Multi->Single for file type '{}' with path '{}'"
+                                     "".format(file_type, input_file))
+        for read in multi_f5.get_reads():
+            try:
+                output_file = os.path.join(output_folder, subfolder, "{}.fast5".format(read.read_id))
+                create_single_f5(output_file, read)
+                output_files.append(os.path.basename(output_file))
+            except Exception as e:
+                logger.error("{}\n\tFailed to copy read '{}' from {}"
+                             "".format(str(e), read.read_id, input_file), exc_info=exc_info)
+    return output_files
 
 
 def create_single_f5(output_file, read):
@@ -87,12 +96,15 @@ def main():
                         help="Folder to output SingleRead fast5 files to")
     parser.add_argument('--recursive', action='store_true',
                         help="Search recursively through folders for MultiRead fast5 files")
+    parser.add_argument('--ignore_symlinks', action='store_true',
+                        help="Ignore symlinks when searching recursively for fast5 files")
     parser.add_argument('-t', '--threads', type=int, default=1, required=False,
                         help="Number of threads to use")
     parser.add_argument('-v', '--version', action='version', version=__version__)
     args = parser.parse_args()
 
-    batch_convert_multi_files_to_single(args.input_path, args.save_path, args.threads, args.recursive)
+    batch_convert_multi_files_to_single(args.input_path, args.save_path, args.threads,
+                                        args.recursive, follow_symlinks=not args.ignore_symlinks)
 
 
 if __name__ == '__main__':