+++ test/test_basecall_1d_tools.py
@@ -1,29 +1,17 @@
 import os
 import sys
-import unittest
 import numpy as np
-from shutil import rmtree
 from ont_fast5_api.fast5_file import Fast5File
 from ont_fast5_api.analysis_tools.basecall_1d import Basecall1DTools
-
-test_data = os.path.join(os.path.dirname(__file__), 'data')
-save_path = os.path.join(os.path.dirname(__file__), 'tmp')
+from test.helpers import TestFast5ApiHelper
 
 py3 = sys.version_info.major == 3
 
 
-class TestBasecall1DTools(unittest.TestCase):
-    def setUp(self):
-        self.save_path = save_path
-        if not os.path.exists(self.save_path):
-            os.makedirs(self.save_path)
-
-    def tearDown(self):
-        if os.path.exists(self.save_path):
-            rmtree(self.save_path)
+class TestBasecall1DTools(TestFast5ApiHelper):
 
     def test_001_put_and_retrieve(self):
-        fname = os.path.join(self.save_path, 'test_file.fast5')
+        fname = self.generate_temp_filename()
         if py3:
             dtypes = [('mean', float),
                       ('start', float),